backend/requirements.txt
supabase
python-dotenv
requests
schedule